#915 Change WebConsole to read API token from ogAdmServer.cfg
This commit changes restfunctions.php to read the API token of
ogAmdServer from ogAdmRepo.cfg to ogAdmServer.cfg. This way it avoids
possible communication errors if ogAdmRepo.cfg and ogAdmServer.cfg are
desynced.

In order to restfuntions.php can read ogAdmServer.cfg this commit
also changes the permissions of the file.