#915 Fix WoL command support in procedures and tasks
Irina reports that the Wake on Lan command does not work in procedures.
The client/computer does not turn on, causing that any of the other
commands are executed.

This bug appears when we adapt the procedures and tasks to work with the
new API REST of the ogAdmServer. It is caused because the client needs
to be on to ask itself for the pending commands, but the WoL command can
not work this way because the client is off.

This commit changes the behaviour of recorreProcedimientos to send the
WoL command without the client has to ask.