#915 Fix WoL command support in queued commands
Irina reports that the Wake on Lan command does not work in queued
commands.

This bug appeared when we fixed duplicated command execution when added
to the queue (ccb862b). It is caused because the client needs to be ON
to ask itself for the queued command, but the WoL command can not work
this way because the client is OFF.

This commit changes the behaviour of gestor_Comandos.php to always send
the WoL command immediately.