#915 Fix EliminarImagenCache command on the WebConsole
Carmelo reports that EliminarImagenCache (Delete image from cache)
command does not work. The image remains in place after the execution
of the command.

This command is actually a wrapper of the command EjecutarScript
(POST /shell) with a different ID. Since gestor_Comandos.php did not
expect a different ID, EliminarImagenCache did not work.

This commit changes the behaviour of gestor_Comandos.php to handle
the EliminarImagenCache ID and treat it as EjecutarScript.