#948 Fix entity (Entidad) creation
Pepe Frias reports that the creation of a new entity does not work. The
WebConsole throws this error:
004 : Comando Error - La sentencia SQL del comando no es correcta.

The WebConsole treats the parameter $ogunit as a string but the Database
treats $ogunit as a tinyint. So, when the WebConsole tries to do the
query the Database throws an error code: 1366.

Error number: 1366; Symbol: ER_TRUNCATED_WRONG_VALUE_FOR_FIELD;
SQLSTATE: HY000

Message: Incorrect %s value: '%s' for column '%s' at row %ld

This commit changes the behaviour of gestor_entidades.php to treat
ogunit as an integer.