#927 Remove opengnsys.cron script
This commit removes the opengnsys cron script that checks if the
ogAdmServer is running or using too much CPU and restarts the process.
Now that the ogAdmServer is more stable this script is not needed. Also,
without this script we are going to know better when the ogAdmServer is
having a problem and get better reports.

In case you observe a crash in ogAdmServer, please, run it under valgrind:

  valgrind --leak-check=full --show-leak-kinds=all --track-origins=yes --verbose --log-file=/tmp/ogadmserver-valgrind-out.log /opt/opengnsys/sbin/ogAdmServer -f /opt/opengnsys/etc/ogAdmServer.cfg

And post your 'ogadmserver-valgrind-out.log' file to the mailing list.

This patch changes the installer and the update scripts too. They do
not configure opengnsys.cron anymore. The update script search and
remove the deprecated files as well.