#974 Fix task edit form
We have seen that if you try to edit a task and assign it to a group of
computers the form do not work. This happens because the string
containing the name of primary key of the table gruposordenadores
(Groups of computers) has a trailing white space.

This commit fixes this bug removing the trailing whitespace.