#999 Fix ogClient session command
Irina reports that booting an OS from ogLive is not working. This is happening
because the bootOS shell script kills ogClient before invoking kexec to start
the OS. However, ogClient runs this shell script that initiates the boot
process, so it is also stopping the boot process.

The ogClient needs to close the connection with ogServer before kexec,
otherwise ogServer keeps reporting via REST API that the client is in ogLive
state / busy state. Because after kexec everything is gone basically.

This patch updates the script to:

1) spawn a process to kill the ogClient in 2 seconds, so the ogClient has time
to reply to the ogServer and to close the connection gracefully.

2) spawn a process to invoke kexec in 3 seconds to boot the OS.

[1] 2ead1639fd54bcf92d56cf7bf89b8d9cf4d6d24e