#1066 #1006 Fix RemotePC IP retrieving on running clients
Error in the REST API response of /ous/:id1/labs/:id2/clients/status
that returns the status of the computers in a classroom. It returns
running PCs with an X.X.X.1 and X.X.X.2 IP address instead of those that
are actually running, which are X.X.X.15 and X.X.X.16. The error cause
is, when constructing the JSON response, a foreach that ignores the
"keys" of the response array.

This commit adds the "keys" to the foreach to iterate also through them.