#1066 #1005 Add support for computer groups in RemotePC
When OpenGnsys is used via RemotePC, a failure occurs when
classrooms with groups of computers are included because the
processing and recursion is not controlled correctly because attempts
are made to add objects to a null array.

This commit adds and populates an array with the computers in the group.

Reviewed-by: Javier Sánchez Parra <​jsanchez@soleta.eu>