#1015 Fix updateBootMode and use it when updating computer info
When a computer info is updated sometimes their related pxe files need
to be updated accordingly.

In order to recreate the file the function createBootMode was being
used inside gestor_ordenadores.

Eg: createBootMode ($cmd, $arranque, $ip, $idioma)

This had some problems because the POST message to gestor_ordenadores
was sometimes lacking the ip field. This field is not necessary because
updateBootMode takes the id of the computer and queries the database for
the ip of the computer associated with it.

Use updateBootMode ($cmd, "idordenador", $idordenador, $idioma) instead.

Fixes related mass updater and "moving pc to another classroom" issues.