#802 UEFI: Add boot entries to last boot order
It's been reported that some UEFI implementations of several
manufacturers do not comply with the standard and behave oddly.

In this case it has been noted that when a BootEntry was created
(with efibootmgr) but not added to the BootOrder this new entry did not
survive a reboot. Thus, UEFI booting of a client from the ogLive menu was not
possible.

Adjust calling ogNvramAddEntry so that it adds the new entry to the
BootOrder, but in last place so network boot, usually first option,
remains in place.