#990 restfunctions: wol: require clients and wol_type params only
Webconsole is directly querying the database to build the payload for
requesting the ogServer a Wake On Lan.

ogServer is expected to query the database for the netmask and mac
parameters. Do not require the client request to have such parameters.

Adjust calls to the wol php method by other OpenGnsys components.

Fixes: a35b7c4 ("#990 Use client broadcast address on WoL")