#1041 Fix poweroffconf script
poweroffconf is checking if the system is idle by doing a pgrep of
processes starting with "/opt/opengnsys/" and ignoring the ones
containing $OGETC or "poweroffconf".

ogclient process always falls under this process lookup thus making
the system seem as always active.

Ignore the ogclient process when checking if the system is not idle.