#995 Fix client speed view inside a room
Before, speed information was not shown in the corresponding computer,
but from first to last element in the table, this did not take into
account for shutdown computers or the position inside the room for a
given computer.

This fix maps ips to speed information so it can later be indexed to
show information correctly.

Fixes: 8634dd3 (#995 Show client link speed if available)